requirements.txt
# Core Streamlit and Web Framework
streamlit>=1.28.0

# File Processing Libraries
pillow>=9.0.0
pandas>=1.5.0
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# PDF Processing
PyMuPDF>=1.23.0
pdf2image>=1.16.0
img2pdf>=0.4.0
reportlab>=4.0.0

# Office Document Processing
python-docx>=0.8.11
python-pptx>=0.6.21

# Image Processing
opencv-python>=4.8.0

# OCR Support (Optional)
pytesseract>=0.3.10

# Compression and Archives
zipfile36>=0.1.3

# Data Processing
numpy>=1.21.0
lxml>=4.9.0

# Utilities
pathlib2>=2.3.0
